+++ js/common/view/TermNode.js
@@ -1,8 +1,7 @@
-// Copyright 2017, University of Colorado Boulder
+// Copyright 2018, University of Colorado Boulder
 
 /**
- * Visual representation of a term.
- * Origin is at the center of the term's icon.
+ * Base type for displaying a term.
  *
  * @author Chris Malley (PixelZoom, Inc.)
  */
@@ -19,13 +18,15 @@ define( function( require ) {
   var TermDragListener = require( 'EQUALITY_EXPLORER/common/view/TermDragListener' );
 
   /**
-   * @param {Term} term - associated model element
-   * @param {TermCreator} termCreator - the creator of term
-   * @param {Plate} plate - the plate that term is associated with
+   * @param {TermCreator} termCreator
+   * @param {Term} term
+   * @param {Plate} plate
+   * @param {Node} contentNode
+   * @param {Node} shadowNode
    * @param {Object} [options]
    * @constructor
    */
-  function TermNode( term, termCreator, plate, options ) {
+  function TermNode( termCreator, term, plate, contentNode, shadowNode, options ) {
 
     var self = this;
 
@@ -35,44 +36,27 @@ define( function( require ) {
     }, options );
 
     // @public (read-only)
-    this.term = term;
+    this.contentNodeSize = new Dimension2( contentNode.width, contentNode.height );
 
-    // @private
-    this.termCreator = termCreator;
-
-    var iconNode = new Node( {
-      children: [ term.icon ], // wrap the icon since we're using scenery DAG feature and need to offset it
-
-      // put origin at the center of the icon
-      centerX: 0,
-      centerY: 0
-    } );
-
-    // shadow, offset from the icon
-    var shadowNode = new Node( {
-      children: [ term.shadow ], // wrap the shadow since we're using scenery DAG feature and need to offset it
-      opacity: 0.4,
-      right: iconNode.right + options.shadowOffset.width,
-      bottom: iconNode.bottom + options.shadowOffset.height,
-      visible: false
-    } );
-
-    // model controls visibility of shadow
-    term.shadowVisibleProperty.link( function( shadowVisible ) {
-      shadowNode.visible = shadowVisible;
-    } );
+    contentNode.centerX = 0;
+    contentNode.centerY = 0;
+    shadowNode.right = contentNode.right + options.shadowOffset.width;
+    shadowNode.bottom = contentNode.bottom + options.shadowOffset.height;
 
     assert && assert( !options.children, 'this type defines its children' );
-    options.children = [ shadowNode, iconNode ];
+    options.children = [ shadowNode, contentNode ];
 
-    var haloRadius = Math.max( term.icon.width, term.icon.height );
+    // @public (read-only)
+    this.term = term;
+
+    var haloRadius = Math.max( contentNode.width, contentNode.height );
 
     // @private {Node|null} halo around the icon
     this.haloNode = null;
     if ( term.haloVisibleProperty ) {
 
       this.haloNode = new HaloNode( haloRadius, {
-        center: iconNode.center,
+        center: contentNode.center,
         visible: false
       } );
       options.children.unshift( this.haloNode );
@@ -94,10 +78,15 @@ define( function( require ) {
 
     // model controls location of this Node
     var locationObserver = function( location ) {
-      self.translation = location;
+      self.center = location;
     };
     term.locationProperty.link( locationObserver ); // unlink required
 
+    // model controls visibility of shadow
+    term.shadowVisibleProperty.link( function( shadowVisible ) {
+      shadowNode.visible = shadowVisible;
+    } );
+    
     // @private removeListener and dispose required
     this.dragListener = new TermDragListener( this, term, termCreator, plate, {
       haloRadius: haloRadius
@@ -106,7 +95,9 @@ define( function( require ) {
 
     // @private
     this.disposeTermNode = function() {
-      term.locationProperty.unlink( locationObserver );
+      if ( term.locationProperty.hasListener( locationObserver ) ) {
+        term.locationProperty.unlink( locationObserver );
+      }
       self.removeInputListener( self.dragListener );
       self.dragListener.dispose();
     };
@@ -117,23 +108,24 @@ define( function( require ) {
   return inherit( Node, TermNode, {
 
     /**
-     * Starts a drag cycle.
-     * The user drags a new term out of a panel below the scale by clicking on an TermCreatorNode.
-     * That action causes TermCreatorNode to instantiate a TermNode.  This function allows
-     * TermCreatorNode to forward the startDrag event to the TermNode that it created.
-     * @param {Event} event
-     */
-    startDrag: function( event ) {
-      this.dragListener.startDrag( event );
-    },
-
-    /**
      * @public
      * @override
      */
     dispose: function() {
       this.disposeTermNode();
       Node.prototype.dispose.call( this );
+    },
+
+    /**
+     * Starts a drag cycle.
+     * The user drags a new term out of a panel below the scale by clicking on an TermCreatorNode.
+     * That action causes TermCreatorNode to instantiate a TermNode.  This function allows
+     * TermCreatorNode to forward the startDrag event to the TermNode that it created.
+     * @param {Event} event
+     * @public
+     */
+    startDrag: function( event ) {
+      this.dragListener.startDrag( event );
     }
   } );
 } );
